Move code related to PyQT into gui modules
